requirements.txt
Flask==2.3.2
tensorflow==2.9.2
numpy==1.23.5
Pillow==9.4.0
typing_extensions==3.10.0.2